requirements.txt
# MyOpenClaw runtime dependencies

# CLI and terminal UX
typer>=0.12.0
questionary>=2.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0

# Environment management
python-dotenv>=1.0.0

# Agent framework
langchain>=0.2.0
langchain_core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Provider SDKs
openai>=1.0.0
anthropic>=0.18.0

# Storage and schema
aiosqlite>=0.19.0
pydantic>=2.0.0
